requirements.txt
fastapi
uvicorn
transformers
torch
scikit-learn
pandas
numpy
sentence-transformers
kafka-python
python-multipart
pydantic
requests